Add changes requested by @rmohr
Since the Template JSON generator was already implemented in the
'tools/vms-generator/vms-generator.go' and we want to specifically test
the example Templates generated by this generator, it makes no sense to
maintain almost identical code within 'tests/template_test.go'.  The
same applies to the 'tests/types.go' containing the 'Template' type
definition that was duplicated in 'tools/vms-generator/types.go'.

In order to implement the requested change I needed to separate
functions within tools/vms-generator/vms-generator.go' into separate
module to make it importable outside the scope of the main module of the
vms-generator tool.  The following importable was added to the project:

  "kubevirt.io/kubevirt/tools/vms-generator/utils"

For the purpose of the Template test, I needed to customize the Docker
tag of the Fedora Template, as well as I had to tweak disks and PVCs of
the RHEL Template to make it work within the CI environment.  I added
following functions to 'tools/vmi-generator/utils/utils.go' to fulfill
this requirement:

  GetTestTemplateFedora() *Template
  newTemplateFedoraWithdockerTag(dockerTag string) *Template

  GetTestTemplateRHEL7() *Template
  newTemplateForRHEL7VM(vm *v1.VirtualMachine) *Template

  addEphemeralPVCDisk(spec *v1.VirtualMachineInstanceSpec,
    claimName string, bus string, diskName string, volumeName string)
    *v1.VirtualMachineInstanceSpec

The functions GetTestTemplateFedora() and GetTestTemplateRHEL7() are
being called from the 'tests/template_test.go' and these two function
calls ultimately replaced most of the helper functions previously
defined in the file.

This commit also introduces a set of minor grooming and refactoring
changes to 'tests/template_test.go'.
